Show relative timestamp using mixin
